Making sodium a double type in Summary class

+++ MyHealth.Fitbit.Nutrition/MyHealth.Fitbit.Nutrition/Models/Summary.cs
@@ -10,7 +10,7 @@ namespace MyHealth.Fitbit.Nutrition.Models
         public double fat { get; set; }
         public double fiber { get; set; }
         public double protein { get; set; }
-        public int sodium { get; set; }
+        public double sodium { get; set; }
         public int water { get; set; }
     }
 }
